intersection-observer/scroll-margin-nested-3.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
#scroller { width: 100px; height: 100px; overflow: scroll; }
#scroller2 { width: 130px; height: 130px; overflow: scroll; }
#scroller3 { width: 160px; height: 160px; overflow: scroll; }
#spacer { width: 50px; height: 100px; }
#target { width: 50px; height: 50px; background-color: red; }
</style>

<div id=scroller3>
  <div id=scroller2>
    <div id=scroller>
      <div id=spacer></div>
      <div id=target></div>
    </div>
  </div>
</div>

<script>
let entries = [];

window.onload = function() {
  runTestCycle(testIntersection, "Test scroll margin intersection");

  const observer = new IntersectionObserver(
    es => entries = entries.concat(es),
    {
      scrollMargin: "20px" // Also accounts for scroll bar height
    }
  );
  observer.observe(target);
};

function testIntersection() {
  assert_equals(entries.length, 1, "IntersectionObserverEntryCount");
  assert_true(entries[0].isIntersecting, "isIntersecting");
}
</script>
